+++ src/alt/store/StoreMixin.js
@@ -1,6 +1,4 @@
-import Symbol from 'es-symbol'
-
-import * as Sym from '../symbols/symbols'
+import transmitter from 'transmitter'
 import * as fn from '../../utils/functions'
 
 const StoreMixin = {
@@ -38,7 +36,7 @@ const StoreMixin = {
 
       const validHandlers = ['success', 'error', 'loading']
       validHandlers.forEach((handler) => {
-        if (spec[handler] && !spec[handler][Sym.ACTION_KEY]) {
+        if (spec[handler] && !spec[handler].id) {
           throw new Error(`${handler} handler must be an action function`)
         }
       })
@@ -91,7 +89,7 @@ const StoreMixin = {
         throw new TypeError('exportPublicMethods expects a function')
       }
 
-      this[Sym.PUBLIC_METHODS][methodName] = value
+      this.publicMethods[methodName] = value
     }, [methods])
   },
 
@@ -100,10 +98,10 @@ const StoreMixin = {
   },
 
   on(lifecycleEvent, handler) {
-    if (lifecycleEvent === 'error') {
-      this[Sym.HANDLING_ERRORS] = true
-    }
-    this[Sym.LIFECYCLE].on(lifecycleEvent, handler.bind(this))
+    if (lifecycleEvent === 'error') this.handlesOwnErrors = true
+    const bus = this.lifecycleEvents[lifecycleEvent] || transmitter()
+    this.lifecycleEvents[lifecycleEvent] = bus
+    return bus.subscribe(handler.bind(this))
   },
 
   bindAction(symbol, handler) {
@@ -116,17 +114,17 @@ const StoreMixin = {
 
     if (handler.length > 1) {
       throw new TypeError(
-        `Action handler in store ${this._storeName} for ` +
-        `${(symbol[Sym.ACTION_KEY] || symbol).toString()} was defined with ` +
+        `Action handler in store ${this.displayName} for ` +
+        `${(symbol.id || symbol).toString()} was defined with ` +
         `two parameters. Only a single parameter is passed through the ` +
         `dispatcher, did you mean to pass in an Object instead?`
       )
     }
 
     // You can pass in the constant or the function itself
-    const key = symbol[Sym.ACTION_KEY] ? symbol[Sym.ACTION_KEY] : symbol
-    this[Sym.LISTENERS][key] = handler.bind(this)
-    this[Sym.ALL_LISTENERS].push(Symbol.keyFor(key))
+    const key = symbol.id ? symbol.id : symbol
+    this.actionListeners[key] = handler.bind(this)
+    this.boundListeners.push(key)
   },
 
   bindActions(actions) {
@@ -163,7 +161,7 @@ const StoreMixin = {
 
       if (!listener) {
         throw new ReferenceError(
-          `${methodName} defined but does not exist in ${this._storeName}`
+          `${methodName} defined but does not exist in ${this.displayName}`
         )
       }