+++ test/bound-listeners-test.js
@@ -1,6 +1,5 @@
 import Alt from '../dist/alt-with-runtime'
 import { assert } from 'chai'
-import Symbol from 'es-symbol'
 
 const alt = new Alt()
 
@@ -53,22 +52,41 @@ export default {
     'when using bindAction'() {
       const myStore = alt.createStore(OneAction)
       assert(myStore.boundListeners.length === 1)
-      assert(myStore.boundListeners[0] === Symbol.keyFor(Actions.ONE))
+      assert(myStore.boundListeners[0] === Actions.ONE)
     },
 
     'when using bindListeners'() {
       const myStore = alt.createStore(TwoAction)
       assert(myStore.boundListeners.length === 1)
-      assert(myStore.boundListeners[0] === Symbol.keyFor(Actions.TWO))
+      assert(myStore.boundListeners[0] === Actions.TWO)
     },
 
     'when using bindActions'() {
       const myStore = alt.createStore(BindActions)
       assert(myStore.boundListeners.length === 2)
       assert(
-        myStore.boundListeners.indexOf(Symbol.keyFor(Actions.ONE)) > -1 &&
-        myStore.boundListeners.indexOf(Symbol.keyFor(Actions.TWO)) > -1
+        myStore.boundListeners.indexOf(Actions.ONE) > -1 &&
+        myStore.boundListeners.indexOf(Actions.TWO) > -1
       )
+    },
+
+    'dispatching actions'() {
+      const alt = new Alt()
+
+      const one = alt.generateActions('one')
+      const two = alt.generateActions('one')
+
+      const store = alt.createStore(function Store() {
+        this.bindAction(one.one, function (x) {
+          assert(x === 1)
+        })
+        this.bindAction(two.one, function (x) {
+          assert(x === 2)
+        })
+      })
+
+      alt.dispatch('global.one', 1)
+      alt.dispatch('global.one1', 2)
     },
   }
 }